pyproject.toml
[tool]
[tool.poetry]
name = "html_formatter"
version = "0.1.0"
description = "Experimental HTML formatter"
license = "MIT license"
keywords = ["html_formatter"]
classifiers = ["Development Status :: 2 - Pre-Alpha", "Intended Audience :: Developers", "License :: OSI Approved :: MIT License", "Natural Language :: English", "Programming Language :: Python :: 3", "Programming Language :: Python :: 3.6", "Programming Language :: Python :: 3.7", "Programming Language :: Python :: 3.8"]
homepage = "https://github.com/benhowes/html_formatter"
authors = ["Ben Howes <ben@ben-howes.co.uk>"]
readme = "README.rst"

[tool.poetry.scripts]
html-formatter = "html_formatter.cli:main"

[tool.poetry.dependencies]
python = ">=3.6.1"
click = ">=7.0"
antlr4-python3-runtime = "==4.7.2"
pathspec = "^0.8.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.2.0"
black = "^19.10b0"
isort = "^4.3.21"
bump2version = "==0.5.11"
watchdog = "==0.10.2"
flake8 = "==3.7.9"
tox = "==3.14.6"
coverage = "*"
Sphinx = "==1.8.5"
pytest = "5.4.1"
ipdb = "^0.13.2"


[tool.black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | setup
  | vsot/antlr
)/
'''

[tool.vsot]
indent_size = 2
